+++ module/segment/runtime/spec.go
@@ -26,7 +26,7 @@ func OpenAPISpec() *openapi3.T {
 
 	return &openapi3.T{
 		OpenAPI: "3.0.3",
-		Info:    &openapi3.Info{Title: "Segment API", Version: "2.0.7"},
+		Info:    &openapi3.Info{Title: "Segment API", Version: "2.0.8"},
 		Paths: openapi3.NewPaths(
 			openapi3.WithPath("/segments/preview/count", &openapi3.PathItem{Post: previewCountOperation()}),
 			openapi3.WithPath("/segments", &openapi3.PathItem{Post: createOperation(), Get: listOperation()}),
@@ -146,6 +146,7 @@ func baseOperation(id string, summary string) *openapi3.Operation {
 func segmentSchema() *openapi3.Schema {
 	filterSchema := openapi3.NewObjectSchema().
 		WithProperty("type", openapi3.NewStringSchema()).
+		WithProperty("exclude", openapi3.NewBoolSchema()).
 		WithProperty("value", openapi3.NewObjectSchema()).
 		WithProperty("parameters", openapi3.NewObjectSchema())